requirements.txt
streamlit==1.32.2

numpy==1.26.4
scipy==1.11.4

librosa==0.10.1
soundfile==0.12.1

openai-whisper==20231117

torch==2.1.2
torchaudio==2.1.2

transformers==4.37.2

opencv-python-headless==4.9.0.80
ffmpeg-python==0.2.0